Add test cases for logging functionalities
testLogFileCreated:
Checks if the log file is created.

testInvalidApiKeyEventLog:
Checks if invalid apikey is printed in log and also covers the event error log functionality.

testInfoLogs:
Checks if system info logs and event info logs are printed in log file.
Checks if no trace and debug logs are printed when level is info.

testDebugLogs:
Checks if system debug logs and event debug logs are printed in log file.
Checks if no trace logs are printed when level is debug.

testTraceEventLog:
Checks if event trace logs are printed in log file.