Fix bug 147579179 and implement synchronizerMode configuration
Below is an example of yaml configuration:

edge_config:
   synchronizerMode: 1

If set to 0, the EMG instance should work as is. Non Synchronizer mode.

If set to 1, the EMG instance should work only as a Synchronizer. This mode should not create any worker nodes.
Write responses of bootstrap, products and jwt_public_key to RedisDB at every config_change_poll_interval.